AYS-523 | `auth.util.exception` Package Has Been Moved to `auth` Package
