+++ src/main/java/org/ays/auth/exception/AysUserNotPassiveException.java
@@ -1,4 +1,4 @@
-package org.ays.auth.util.exception;
+package org.ays.auth.exception;
 
 import org.ays.auth.model.enums.AysUserStatus;
 import org.ays.common.exception.AysNotExistException;
