+++ src/main/java/org/ays/auth/service/AysUserPasswordService.java
@@ -1,10 +1,10 @@
 package org.ays.auth.service;
 
+import org.ays.auth.exception.AysEmailAddressNotValidException;
+import org.ays.auth.exception.AysUserPasswordCannotChangedException;
+import org.ays.auth.exception.AysUserPasswordDoesNotExistException;
 import org.ays.auth.model.request.AysPasswordCreateRequest;
 import org.ays.auth.model.request.AysPasswordForgotRequest;
-import org.ays.auth.util.exception.AysEmailAddressNotValidException;
-import org.ays.auth.util.exception.AysUserPasswordCannotChangedException;
-import org.ays.auth.util.exception.AysUserPasswordDoesNotExistException;
 
 /**
  * Service interface for handling user password operations.
