+++ src/main/java/org/ays/auth/service/impl/AysAuthServiceImpl.java
@@ -2,6 +2,11 @@ package org.ays.auth.service.impl;
 
 import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
+import org.ays.auth.exception.AysEmailAddressNotValidException;
+import org.ays.auth.exception.AysPasswordNotValidException;
+import org.ays.auth.exception.AysUserDoesNotAccessPageException;
+import org.ays.auth.exception.AysUserIdNotValidException;
+import org.ays.auth.exception.AysUserNotActiveException;
 import org.ays.auth.model.AysIdentity;
 import org.ays.auth.model.AysRole;
 import org.ays.auth.model.AysToken;
@@ -14,11 +19,6 @@ import org.ays.auth.port.AysUserSavePort;
 import org.ays.auth.service.AysAuthService;
 import org.ays.auth.service.AysInvalidTokenService;
 import org.ays.auth.service.AysTokenService;
-import org.ays.auth.util.exception.AysEmailAddressNotValidException;
-import org.ays.auth.util.exception.AysPasswordNotValidException;
-import org.ays.auth.util.exception.AysUserDoesNotAccessPageException;
-import org.ays.auth.util.exception.AysUserIdNotValidException;
-import org.ays.auth.util.exception.AysUserNotActiveException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
