+++ src/main/java/org/ays/auth/service/impl/AysRoleUpdateServiceImpl.java
@@ -1,6 +1,13 @@
 package org.ays.auth.service.impl;
 
 import lombok.RequiredArgsConstructor;
+import org.ays.auth.exception.AysInvalidRoleStatusException;
+import org.ays.auth.exception.AysPermissionNotExistException;
+import org.ays.auth.exception.AysRoleAlreadyDeletedException;
+import org.ays.auth.exception.AysRoleAlreadyExistsByNameException;
+import org.ays.auth.exception.AysRoleAssignedToUserException;
+import org.ays.auth.exception.AysRoleNotExistByIdException;
+import org.ays.auth.exception.AysUserNotSuperAdminException;
 import org.ays.auth.model.AysIdentity;
 import org.ays.auth.model.AysPermission;
 import org.ays.auth.model.AysRole;
@@ -10,13 +17,6 @@ import org.ays.auth.port.AysPermissionReadPort;
 import org.ays.auth.port.AysRoleReadPort;
 import org.ays.auth.port.AysRoleSavePort;
 import org.ays.auth.service.AysRoleUpdateService;
-import org.ays.auth.util.exception.AysInvalidRoleStatusException;
-import org.ays.auth.util.exception.AysPermissionNotExistException;
-import org.ays.auth.util.exception.AysRoleAlreadyDeletedException;
-import org.ays.auth.util.exception.AysRoleAlreadyExistsByNameException;
-import org.ays.auth.util.exception.AysRoleAssignedToUserException;
-import org.ays.auth.util.exception.AysRoleNotExistByIdException;
-import org.ays.auth.util.exception.AysUserNotSuperAdminException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
