+++ src/main/java/org/ays/auth/service/impl/AysTokenServiceImpl.java
@@ -13,10 +13,10 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.time.DateUtils;
 import org.ays.auth.config.AysTokenConfigurationParameter;
+import org.ays.auth.exception.AysTokenNotValidException;
 import org.ays.auth.model.AysToken;
 import org.ays.auth.model.enums.AysTokenClaims;
 import org.ays.auth.service.AysTokenService;
-import org.ays.auth.util.exception.AysTokenNotValidException;
 import org.ays.common.util.AysListUtil;
 import org.ays.common.util.AysRandomUtil;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
