+++ src/main/java/org/ays/auth/service/impl/AysUserPasswordServiceImpl.java
@@ -1,6 +1,9 @@
 package org.ays.auth.service.impl;
 
 import lombok.RequiredArgsConstructor;
+import org.ays.auth.exception.AysEmailAddressNotValidException;
+import org.ays.auth.exception.AysUserPasswordCannotChangedException;
+import org.ays.auth.exception.AysUserPasswordDoesNotExistException;
 import org.ays.auth.model.AysUser;
 import org.ays.auth.model.request.AysPasswordCreateRequest;
 import org.ays.auth.model.request.AysPasswordForgotRequest;
@@ -8,9 +11,6 @@ import org.ays.auth.port.AysUserReadPort;
 import org.ays.auth.port.AysUserSavePort;
 import org.ays.auth.service.AysUserMailService;
 import org.ays.auth.service.AysUserPasswordService;
-import org.ays.auth.util.exception.AysEmailAddressNotValidException;
-import org.ays.auth.util.exception.AysUserPasswordCannotChangedException;
-import org.ays.auth.util.exception.AysUserPasswordDoesNotExistException;
 import org.ays.common.util.AysRandomUtil;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
