+++ src/test/java/org/ays/auth/service/impl/AysAuthServiceImplTest.java
@@ -2,6 +2,13 @@ package org.ays.auth.service.impl;
 
 import io.jsonwebtoken.Claims;
 import org.ays.AysUnitTest;
+import org.ays.auth.exception.AysEmailAddressNotValidException;
+import org.ays.auth.exception.AysPasswordNotValidException;
+import org.ays.auth.exception.AysTokenAlreadyInvalidatedException;
+import org.ays.auth.exception.AysTokenNotValidException;
+import org.ays.auth.exception.AysUserDoesNotAccessPageException;
+import org.ays.auth.exception.AysUserIdNotValidException;
+import org.ays.auth.exception.AysUserNotActiveException;
 import org.ays.auth.model.AysIdentity;
 import org.ays.auth.model.AysPermission;
 import org.ays.auth.model.AysPermissionBuilder;
@@ -20,13 +27,6 @@ import org.ays.auth.port.AysUserReadPort;
 import org.ays.auth.port.AysUserSavePort;
 import org.ays.auth.service.AysInvalidTokenService;
 import org.ays.auth.service.AysTokenService;
-import org.ays.auth.util.exception.AysEmailAddressNotValidException;
-import org.ays.auth.util.exception.AysPasswordNotValidException;
-import org.ays.auth.util.exception.AysTokenAlreadyInvalidatedException;
-import org.ays.auth.util.exception.AysTokenNotValidException;
-import org.ays.auth.util.exception.AysUserDoesNotAccessPageException;
-import org.ays.auth.util.exception.AysUserIdNotValidException;
-import org.ays.auth.util.exception.AysUserNotActiveException;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
