+++ src/test/java/org/ays/auth/service/impl/AysRoleCreateServiceImplTest.java
@@ -1,6 +1,9 @@
 package org.ays.auth.service.impl;
 
 import org.ays.AysUnitTest;
+import org.ays.auth.exception.AysPermissionNotExistException;
+import org.ays.auth.exception.AysRoleAlreadyExistsByNameException;
+import org.ays.auth.exception.AysUserNotSuperAdminException;
 import org.ays.auth.model.AysIdentity;
 import org.ays.auth.model.AysPermission;
 import org.ays.auth.model.AysPermissionBuilder;
@@ -10,9 +13,6 @@ import org.ays.auth.model.request.AysRoleCreateRequestBuilder;
 import org.ays.auth.port.AysPermissionReadPort;
 import org.ays.auth.port.AysRoleReadPort;
 import org.ays.auth.port.AysRoleSavePort;
-import org.ays.auth.util.exception.AysPermissionNotExistException;
-import org.ays.auth.util.exception.AysRoleAlreadyExistsByNameException;
-import org.ays.auth.util.exception.AysUserNotSuperAdminException;
 import org.ays.common.util.AysRandomUtil;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
