+++ src/test/java/org/ays/auth/service/impl/AysTokenServiceImplTest.java
@@ -7,12 +7,12 @@ import io.jsonwebtoken.Jwts;
 import org.apache.commons.lang3.time.DateUtils;
 import org.ays.AysUnitTest;
 import org.ays.auth.config.AysTokenConfigurationParameter;
+import org.ays.auth.exception.AysTokenNotValidException;
 import org.ays.auth.model.AysToken;
 import org.ays.auth.model.AysUser;
 import org.ays.auth.model.AysUserBuilder;
 import org.ays.auth.model.enums.AysTokenClaims;
 import org.ays.auth.util.AysKeyConverter;
-import org.ays.auth.util.exception.AysTokenNotValidException;
 import org.ays.common.util.AysListUtil;
 import org.ays.common.util.AysRandomUtil;
 import org.junit.jupiter.api.Assertions;
