+++ src/test/java/org/ays/auth/service/impl/AysUserReadServiceImplTest.java
@@ -1,6 +1,7 @@
 package org.ays.auth.service.impl;
 
 import org.ays.AysUnitTest;
+import org.ays.auth.exception.AysUserNotExistByIdException;
 import org.ays.auth.model.AysIdentity;
 import org.ays.auth.model.AysUser;
 import org.ays.auth.model.AysUserBuilder;
@@ -8,7 +9,6 @@ import org.ays.auth.model.AysUserFilter;
 import org.ays.auth.model.request.AysUserListRequest;
 import org.ays.auth.model.request.AysUserListRequestBuilder;
 import org.ays.auth.port.AysUserReadPort;
-import org.ays.auth.util.exception.AysUserNotExistByIdException;
 import org.ays.common.model.AysPage;
 import org.ays.common.model.AysPageBuilder;
 import org.ays.common.model.AysPageable;
