+++ src/test/java/org/ays/auth/service/impl/AysUserUpdateServiceImplTest.java
@@ -1,6 +1,14 @@
 package org.ays.auth.service.impl;
 
 import org.ays.AysUnitTest;
+import org.ays.auth.exception.AysRolesNotExistException;
+import org.ays.auth.exception.AysUserAlreadyDeletedException;
+import org.ays.auth.exception.AysUserAlreadyExistsByEmailAddressException;
+import org.ays.auth.exception.AysUserAlreadyExistsByPhoneNumberException;
+import org.ays.auth.exception.AysUserIsNotActiveOrPassiveException;
+import org.ays.auth.exception.AysUserNotActiveException;
+import org.ays.auth.exception.AysUserNotExistByIdException;
+import org.ays.auth.exception.AysUserNotPassiveException;
 import org.ays.auth.model.AysIdentity;
 import org.ays.auth.model.AysRole;
 import org.ays.auth.model.AysRoleBuilder;
@@ -12,14 +20,6 @@ import org.ays.auth.model.request.AysUserUpdateRequestBuilder;
 import org.ays.auth.port.AysRoleReadPort;
 import org.ays.auth.port.AysUserReadPort;
 import org.ays.auth.port.AysUserSavePort;
-import org.ays.auth.util.exception.AysRolesNotExistException;
-import org.ays.auth.util.exception.AysUserAlreadyDeletedException;
-import org.ays.auth.util.exception.AysUserAlreadyExistsByEmailAddressException;
-import org.ays.auth.util.exception.AysUserAlreadyExistsByPhoneNumberException;
-import org.ays.auth.util.exception.AysUserIsNotActiveOrPassiveException;
-import org.ays.auth.util.exception.AysUserNotActiveException;
-import org.ays.auth.util.exception.AysUserNotExistByIdException;
-import org.ays.auth.util.exception.AysUserNotPassiveException;
 import org.ays.common.model.AysPhoneNumber;
 import org.ays.common.model.AysPhoneNumberBuilder;
 import org.ays.common.util.AysRandomUtil;
